+++ sktime/markovprocess/_base.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from sktime.base import Estimator
-from sktime.markovprocess._dtraj_stats import blocksplit_dtrajs, cvsplit_dtrajs
+from sktime.markovprocess.transition_counting import blocksplit_dtrajs, cvsplit_dtrajs
 
 
 # TODO: distinguish more between model and estimator attributes.
@@ -146,55 +146,3 @@ def score_cv(estimator: _MSMBaseEstimator, dtrajs, n=10, score_method='VAMP2', s
         scores.append(s)
     return np.array(scores)
 
-
-    # TODO: this one is tricky
-def cktest(test_estimator, test_model, dtrajs, nsets, memberships=None, mlags=10,
-           conf=0.95, err_est=False):
-    """ Conducts a Chapman-Kolmogorow test.
-
-    Parameters
-    ----------
-    nsets : int
-        number of sets to test on
-    memberships : ndarray(nstates, nsets), optional
-        optional state memberships. By default (None) will conduct a cktest
-        on PCCA (metastable) sets.
-    mlags : int or int-array, optional
-        multiples of lag times for testing the Model, e.g. range(10).
-        A single int will trigger a range, i.e. mlags=10 maps to
-        mlags=range(10). The setting None will choose mlags automatically
-        according to the longest available trajectory
-    conf : float, optional
-        confidence interval
-    err_est : bool, optional
-        compute errors also for all estimations (computationally expensive)
-        If False, only the prediction will get error bars, which is often
-        sufficient to validate a model.
-
-    Returns
-    -------
-    cktest : :class:`ChapmanKolmogorovValidator <sktime.markovprocess.ChapmanKolmogorovValidator>`
-
-
-    References
-    ----------
-    This test was suggested in [1]_ and described in detail in [2]_.
-
-    .. [1] F. Noe, Ch. Schuette, E. Vanden-Eijnden, L. Reich and
-        T. Weikl: Constructing the Full Ensemble of Folding Pathways
-        from Short Off-Equilibrium Simulations.
-        Proc. Natl. Acad. Sci. USA, 106, 19011-19016 (2009)
-    .. [2] Prinz, J H, H Wu, M Sarich, B Keller, M Senne, M Held, J D
-        Chodera, C Schuette and F Noe. 2011. Markov models of
-        molecular kinetics: Generation and validation. J Chem Phys
-        134: 174105
-
-    """
-    if memberships is None:
-        test_model.pcca(nsets)
-        memberships = test_model.metastable_memberships
-    from sktime.markovprocess.chapman_kolmogorov_validator import ChapmanKolmogorovValidator
-    ck = ChapmanKolmogorovValidator(test_estimator=test_estimator, test_model=test_model, memberships=memberships,
-                                    mlags=mlags, conf=conf, err_est=err_est)
-    ck.fit(dtrajs)
-    return ck
