+++ sktime/markovprocess/transition_counting.py
